+++ .editorconfig
@@ -8,9 +8,6 @@ indent_style = space
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[*.blade.php]
-indent_size = 2
-
 [*.js]
 indent_size = 2
